% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Cleaning function}
\usage{
clean_data(df, id_var, na_threshold = 0.9, ...)
}
\arguments{
\item{df}{data.frame, data to be cleaned}

\item{id_var}{character, columnname of dataframe which acts as the ID variable}

\item{na_threshold}{numeric, if NAs within a column are greater than the treshold given, this columns will be excluded}

\item{...}{args forwarded to \code{\link{read_csv2}}. Hint: You can edit the import column type by the col_types argument.}
}
\value{
data.frame, excluding columns with too many NAs and impute existing NAs in numeric fields with zero
}
\description{
Data is cleand
}
\examples{
\dontrun{
clean_data(CompanyData2018, id_var = "ID")
}
}
