% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CompanyData2018}
\alias{CompanyData2018}
\title{CompanyData2018 data with 285 variables}
\format{A data frame with 13012 rows and 285 variables

\describe{
  \item{LegalStatusCode}{double}
  \item{Judet}{character}
  \item{NaceFinal}{double}
  \item{Year}{double}
  \item{CompanyRegDate}{character}
  \item{CompanyAge}{double}
  \item{CompanyLastReg}{double}
  \item{NaceCode}{double}
  \item{NaceAnaf2017}{double}
  \item{NaceActualDiff}{double}
  \item{ApartmentCompany}{double}
  \item{OrganizationType}{character}
  \item{IndustrySector}{character}
  \item{IndustryRisk}{character}
  \item{Region}{character}
  \item{OperatingCode}{double}
  \item{InsolvencyFlag}{double}
  \item{CriminalPersecutionFlag}{double}
  \item{MergersFlag}{double}
  \item{AdverseInformationFlag}{double}
  \item{Art15Flag}{double}
  \item{LegalStatus}{character}
  \item{IsFingerPrinciple}{logical}
  \item{CodGroupFinal}{double}
  \item{ID}{integer}
  \item{YearsManagerialExperience}{double}
  \item{OneManShow}{double}
  \item{FamilyCompany}{double}
  \item{GenCountAdminPI}{double}
  \item{GenCountAdminLE}{double}
  \item{GenCountAdminList}{double}
  \item{GenCountOwners}{double}
  \item{GenSumOwnership}{double}
  \item{GenMinOwnership}{double}
  \item{GenMaxOwnership}{double}
  \item{GenAvgOwnership}{double}
  \item{GenMinOwnersStart}{double}
  \item{GenMaxOwnersStart}{double}
  \item{GenAvgOwnersStart}{double}
  \item{GenCountOwnersPI}{double}
  \item{GenCountOwnersLE}{double}
  \item{GenCountOwnersList}{double}
  \item{GenSumOwnershipPI}{double}
  \item{GenSumOwnershipLE}{double}
  \item{GenSumOwnershipList}{double}
  \item{GenMinOwnershipPI}{double}
  \item{GenMinOwnershipLE}{double}
  \item{GenMinOwnershipList}{double}
  \item{GenMaxOwnershipPI}{double}
  \item{GenMaxOwnershipLE}{double}
  \item{GenMaxOwnershipList}{double}
  \item{GenAvgOwnershipPI}{double}
  \item{GenAvgOwnershipLE}{double}
  \item{GenAvgOwnershipList}{double}
  \item{GenMinOwnersStartPI}{double}
  \item{GenMinOwnersStartLE}{double}
  \item{GenMinOwnersStartList}{double}
  \item{GenMaxOwnersStartPI}{double}
  \item{GenMaxOwnersStartLE}{double}
  \item{GenMaxOwnersStartList}{double}
  \item{GenAvgOwnersStartPI}{double}
  \item{GenAvgOwnersStartLE}{double}
  \item{GenAvgOwnersStartList}{double}
  \item{GenMinOwnersAgePI}{double}
  \item{GenMaxOwnersAgePI}{double}
  \item{GenAvgOwnersAgePI}{double}
  \item{GenCountOwnersRes}{double}
  \item{GenSumOwnershipRes}{double}
  \item{GenMinOwnershipRes}{double}
  \item{GenMaxOwnershipRes}{double}
  \item{GenAvgOwnershipRes}{double}
  \item{GenMinOwnersResStart}{double}
  \item{GenMaxOwnersResStart}{double}
  \item{GenAvgOwnersResStart}{double}
  \item{GenCountOwnersPIRes}{double}
  \item{GenCountOwnersLERes}{double}
  \item{GenSumOwnershipPIRes}{double}
  \item{GenSumOwnershipLERes}{double}
  \item{GenMinOwnershipPIRes}{double}
  \item{GenMinOwnershipLERes}{double}
  \item{GenMaxOwnershipPIRes}{double}
  \item{GenMaxOwnershipLERes}{double}
  \item{GenAvgOwnershipPIRes}{double}
  \item{GenAvgOwnershipLERes}{double}
  \item{GenMinOwnersStartPIRes}{double}
  \item{GenMinOwnersStartLERes}{double}
  \item{GenMaxOwnersStartPIRes}{double}
  \item{GenMaxOwnersStartLERes}{double}
  \item{GenAvgOwnersStartPIRes}{double}
  \item{GenAvgOwnersStartLERes}{double}
  \item{GenMinOwnersAgePIRes}{double}
  \item{GenMinOwnersAgeLERes}{logical}
  \item{GenMaxOwnersAgePIRes}{double}
  \item{GenMaxOwnersAgeLERes}{logical}
  \item{GenAvgOwnersAgePIRes}{double}
  \item{GenAvgOwnersAgeLERes}{logical}
  \item{GenCountOwnersNonRes}{double}
  \item{GenSumOwnershipNonRes}{double}
  \item{GenMinOwnershipNonRes}{double}
  \item{GenMaxOwnershipNonRes}{double}
  \item{GenAvgOwnershipNonRes}{double}
  \item{GenMinOwnersNonResStart}{double}
  \item{GenMaxOwnersNonResStart}{double}
  \item{GenAvgOwnersNonResStart}{double}
  \item{GenCountOwnersPINonRes}{double}
  \item{GenCountOwnersLENonRes}{double}
  \item{GenSumOwnershipPINonRes}{double}
  \item{GenSumOwnershipLENonRes}{double}
  \item{GenMinOwnershipPINonRes}{double}
  \item{GenMinOwnershipLENonRes}{double}
  \item{GenMaxOwnershipPINonRes}{double}
  \item{GenMaxOwnershipLENonRes}{double}
  \item{GenAvgOwnershipPINonRes}{double}
  \item{GenAvgOwnershipLENonRes}{double}
  \item{GenMinOwnersStartPINonRes}{double}
  \item{GenMinOwnersStartLENonRes}{double}
  \item{GenMaxOwnersStartPINonRes}{double}
  \item{GenMaxOwnersStartLENonRes}{double}
  \item{GenAvgOwnersStartPINonRes}{double}
  \item{GenAvgOwnersStartLENonRes}{double}
  \item{GenMinOwnersAgePINonRes}{double}
  \item{GenMinOwnersAgeLENonRes}{logical}
  \item{GenMaxOwnersAgePINonRes}{double}
  \item{GenMaxOwnersAgeLENonRes}{logical}
  \item{GenAvgOwnersAgePINonRes}{double}
  \item{GenAvgOwnersAgeLENonRes}{logical}
  \item{GenCountAdmin}{double}
  \item{GenMinAdminStart}{double}
  \item{GenMaxAdminStart}{double}
  \item{GenAvgAdminStart}{double}
  \item{GenMinAdminStartPI}{double}
  \item{GenMinAdminStartLE}{logical}
  \item{GenMinAdminStartList}{logical}
  \item{GenMaxAdminStartPI}{double}
  \item{GenMaxAdminStartLE}{logical}
  \item{GenMaxAdminStartList}{logical}
  \item{GenAvgAdminStartPI}{double}
  \item{GenAvgAdminStartLE}{logical}
  \item{GenAvgAdminStartList}{logical}
  \item{GenMinAdminAgePI}{double}
  \item{GenMaxAdminAgePI}{double}
  \item{GenAvgAdminAgePI}{double}
  \item{GenCountAdminRes}{double}
  \item{GenMinAdminResStart}{double}
  \item{GenMaxAdminResStart}{double}
  \item{GenAvgAdminResStart}{double}
  \item{GenCountAdminPIRes}{double}
  \item{GenCountAdminLERes}{double}
  \item{GenMinAdminStartPIRes}{double}
  \item{GenMinAdminStartLERes}{logical}
  \item{GenMaxAdminStartPIRes}{double}
  \item{GenMaxAdminStartLERes}{logical}
  \item{GenAvgAdminStartPIRes}{double}
  \item{GenAvgAdminStartLERes}{logical}
  \item{GenMinAdminAgePIRes}{double}
  \item{GenMaxAdminAgePIRes}{double}
  \item{GenAvgAdminAgePIRes}{double}
  \item{GenCountAdminNonRes}{double}
  \item{GenMinAdminNonResStart}{double}
  \item{GenMaxAdminNonResStart}{double}
  \item{GenAvgAdminNonResStart}{double}
  \item{GenCountAdminPINonRes}{double}
  \item{GenCountAdminLENonRes}{double}
  \item{GenMinAdminStartPINonRes}{double}
  \item{GenMinAdminStartLENonRes}{logical}
  \item{GenMaxAdminStartPINonRes}{double}
  \item{GenMaxAdminStartLENonRes}{logical}
  \item{GenAvgAdminStartPINonRes}{double}
  \item{GenAvgAdminStartLENonRes}{logical}
  \item{GenMinAdminAgePINonRes}{double}
  \item{GenMaxAdminAgePINonRes}{double}
  \item{GenAvgAdminAgePINonRes}{double}
  \item{GenCountAuditor}{logical}
  \item{GenMinAuditorStart}{logical}
  \item{GenMaxAuditorStart}{logical}
  \item{GenAvgAuditorStart}{logical}
  \item{GenMinAuditorStartPI}{logical}
  \item{GenMinAuditorStartLE}{logical}
  \item{GenMinAuditorStartList}{logical}
  \item{GenMaxAuditorStartPI}{logical}
  \item{GenMaxAuditorStartLE}{logical}
  \item{GenMaxAuditorStartList}{logical}
  \item{GenAvgAuditorStartPI}{logical}
  \item{GenAvgAuditorStartLE}{logical}
  \item{GenAvgAuditorStartList}{logical}
  \item{GenCountFormerOwners}{logical}
  \item{GenSumFormerOwnership}{logical}
  \item{GenMinFormerOwnership}{logical}
  \item{GenMaxFormerOwnership}{logical}
  \item{GenAvgFormerOwnership}{logical}
  \item{GenMinFormerOwnersEnd}{logical}
  \item{GenMaxOwnersEnd}{logical}
  \item{GenAvgOwnersEnd}{logical}
  \item{GenCountFormerOwnersPI}{logical}
  \item{GenCountFormerOwnersLE}{logical}
  \item{GenCountFormerOwnersList}{logical}
  \item{GenSumFormerOwnershipPI}{logical}
  \item{GenSumFormerOwnershipLE}{logical}
  \item{GenSumFormerOwnershipList}{logical}
  \item{GenMinFormerOwnershipPI}{logical}
  \item{GenMinFormerOwnershipLE}{logical}
  \item{GenMinFormerOwnershipList}{logical}
  \item{GenMaxFormerOwnershipPI}{logical}
  \item{GenMaxFormerOwnershipLE}{logical}
  \item{GenMaxFormerOwnershipList}{logical}
  \item{GenAvgFormerOwnershipPI}{logical}
  \item{GenAvgFormerOwnershipLE}{logical}
  \item{GenAvgFormerOwnershipList}{logical}
  \item{GenMinFormerOwnersEndPI}{logical}
  \item{GenMinFormerOwnersEndLE}{logical}
  \item{GenMinFormerOwnersEndList}{logical}
  \item{GenMaxFormerOwnersEndPI}{logical}
  \item{GenMaxFormerOwnersEndLE}{logical}
  \item{GenMaxFormerOwnersEndList}{logical}
  \item{GenAvgFormerOwnersEndPI}{logical}
  \item{GenAvgFormerOwnersEndLE}{logical}
  \item{GenAvgFormerOwnersEndList}{logical}
  \item{GenMinFormerOwnersAgePI}{logical}
  \item{GenMaxFormerOwnersAgePI}{logical}
  \item{GenAvgFormerOwnersAgePI}{logical}
  \item{GenCountFormerAdmin}{double}
  \item{GenMinFormerAdminEnd}{double}
  \item{GenMaxFormerAdminEnd}{double}
  \item{GenAvgFormerAdminEnd}{double}
  \item{GenCountFormerAdminPI}{double}
  \item{GenCountFormerAdminLE}{double}
  \item{GenCountFormerAdminList}{double}
  \item{GenMinFormerAdminEndPI}{double}
  \item{GenMinFormerAdminEndLE}{double}
  \item{GenMinFormerAdminEndList}{logical}
  \item{GenMaxFormerAdminEndPI}{double}
  \item{GenMaxFormerAdminEndLE}{double}
  \item{GenMaxFormerAdminEndList}{logical}
  \item{GenAvgFormerAdminEndPI}{double}
  \item{GenAvgFormerAdminEndLE}{double}
  \item{GenAvgFormerAdminEndList}{logical}
  \item{GenMinAFormerdminAgePI}{double}
  \item{GenMaxFormerAdminAgePI}{double}
  \item{GenAvgFormerAdminAgePI}{double}
  \item{GenCountFormerAuditor}{double}
  \item{GenMinFormerAuditorEnd}{double}
  \item{GenMaxFormerAuditorEnd}{double}
  \item{GenAvgFormerAuditorEnd}{double}
  \item{GenCountFormerAuditorPI}{double}
  \item{GenCountFormerAuditorLE}{double}
  \item{GenCountFormerAuditorList}{double}
  \item{GenMinFormerAuditorEndPI}{double}
  \item{GenMinFormerAuditorEndLE}{double}
  \item{GenMinFormerAuditorEndList}{logical}
  \item{GenMaxFormerAuditorEndPI}{double}
  \item{GenMaxFormerAuditorEndLE}{double}
  \item{GenMaxFormerAuditorEndList}{logical}
  \item{GenAvgFormerAuditorEndPI}{double}
  \item{GenAvgFormerAuditorEndLE}{double}
  \item{GenAvgFormerAuditorEndList}{logical}
  \item{OwnersPerChangeNew12M}{double}
  \item{OwnersPerChangeNew24M}{double}
  \item{OwnersPerChangeNew36M}{double}
  \item{OwnersNoChangeNew12M}{double}
  \item{OwnersNoChangeNew24M}{double}
  \item{OwnersNoChangeNew36M}{double}
  \item{AdminNoChangeNew12M}{double}
  \item{AdminNoChangeNew24M}{double}
  \item{AdminNoChangeNew36M}{double}
  \item{AuditorNoChangeNew12M}{double}
  \item{AuditorNoChangeNew24M}{double}
  \item{AuditorNoChangeNew36M}{double}
  \item{OwnersResPerChangeNew12M}{double}
  \item{OwnersResPerChangeNew24M}{double}
  \item{OwnersResPerChangeNew36M}{double}
  \item{OwnersResNoChangeNew12M}{double}
  \item{OwnersResNoChangeNew24M}{double}
  \item{OwnersResNoChangeNew36M}{double}
  \item{AdminResNoChangeNew12M}{double}
  \item{AdminResNoChangeNew24M}{double}
  \item{AdminResNoChangeNew36M}{double}
  \item{OwnersNonResPerChangeNew12M}{double}
  \item{OwnersNonResPerChangeNew24M}{double}
  \item{OwnersNonResPerChangeNew36M}{double}
  \item{OwnersNonResNoChangeNew12M}{double}
  \item{OwnersNonResNoChangeNew24M}{double}
  \item{OwnersNonResNoChangeNew36M}{double}
  \item{AdminNonResNoChangeNew12M}{double}
  \item{AdminNonResNoChangeNew24M}{double}
  \item{AdminNonResNoChangeNew36M}{double}
}}
\usage{
CompanyData2018
}
\description{
CompanyData2018 data with 285 variables
}
\keyword{datasets}
