% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Financials2018}
\alias{Financials2018}
\title{Financials 2018 data with 165 variables}
\format{A data frame with 1240 rows and 165 variables

\describe{
  \item{ID}{integer}
  \item{TotalSales}{double}
  \item{TotalNetSales}{double}
  \item{TotalOtherIncome}{double}
  \item{OwnWorkCapitalised}{double}
  \item{ChangeInventoryProd}{double}
  \item{OtherOperatingIncome}{double}
  \item{iOOIProvisionRelease}{double}
  \item{iOOISubsidies}{double}
  \item{iOOISubsidiesExtraord}{double}
  \item{iOOIExtraordinary}{double}
  \item{TotalRevenues}{double}
  \item{Cogs}{double}
  \item{iCogsRawMaterials}{double}
  \item{iCogsMerchandise}{double}
  \item{iCogsSubdeliveries}{double}
  \item{PersonnelExpenses}{double}
  \item{SGAOtherOpExpenses}{double}
  \item{iSGABadDebtProv}{double}
  \item{iSGAExtraordinary}{double}
  \item{DepreciationAndAmortization}{double}
  \item{Ebit}{double}
  \item{Ebitda}{double}
  \item{EbitdaMargin}{character}
  \item{TradingMargin}{character}
  \item{ProductionMargin}{double}
  \item{InterestIncome}{double}
  \item{InterestExpense}{double}
  \item{NetInterestResult}{double}
  \item{FXGains}{double}
  \item{FXLosses}{double}
  \item{NetFXResult}{double}
  \item{GroupInvestResult}{double}
  \item{iSubsidiarySales}{double}
  \item{NetFinancialResult}{double}
  \item{ExtraordinaryIncome}{double}
  \item{iExtraordFixedAssetsReval}{double}
  \item{iExtraordProvRelease}{double}
  \item{ExtraordinaryExpenses}{double}
  \item{iExtraordFAWriteDowns}{double}
  \item{iExtraordProvTransfer}{double}
  \item{NetExtraordinaryResult}{double}
  \item{IncomeTaxes}{double}
  \item{NetProfitLoss}{double}
  \item{CashBankDeposits}{double}
  \item{STSecurities}{double}
  \item{OtherSTFinAssets}{double}
  \item{CashAndCashEquivalents}{double}
  \item{TradeReceivables}{double}
  \item{IntragroupReceivables}{double}
  \item{ReceivablesTaxAuthority}{double}
  \item{OtherReceivables}{double}
  \item{iOtherReceivablesOwners}{double}
  \item{TotalReceivables}{double}
  \item{RawMaterials}{double}
  \item{Merchandise}{double}
  \item{TotalInventory}{double}
  \item{PrepaidExpenses}{double}
  \item{TotalCurrentAssets}{double}
  \item{LandBuildings}{double}
  \item{PlantMachinery}{double}
  \item{FurnitureEquipment}{double}
  \item{AdvancesTangibleAssets}{double}
  \item{TangibleAssetsInProgress}{double}
  \item{TangibleFixedAssets}{double}
  \item{RightsPatents}{double}
  \item{OtherCapitalizedExpenses}{double}
  \item{IntangibleFixedAssets}{double}
  \item{OtherLTFinancialAssets}{double}
  \item{TotalFixedAssets}{double}
  \item{TotalAssets}{double}
  \item{STBankLoans}{double}
  \item{OtherSTLoansNonBank}{double}
  \item{SubtotalSTInterestBearingDebt}{double}
  \item{STAccountsPayables}{double}
  \item{STIntragroupPayables}{double}
  \item{STProvisions}{double}
  \item{AccruedExpenses}{double}
  \item{PayablesTaxAuthority}{double}
  \item{DividendsPayable}{double}
  \item{OtherSTLiabilities}{double}
  \item{TotalSTLiabilities}{double}
  \item{LTBankLoans}{double}
  \item{SHLoansGranted}{double}
  \item{CapitalizedLeaseObligations}{double}
  \item{SubtotalLTInterestBearingDebt}{double}
  \item{OtherLTLiabilities}{double}
  \item{TotalLTLiabilities}{double}
  \item{TotalLiabilities}{double}
  \item{SeniorDebt}{double}
  \item{ShareCapital}{double}
  \item{SharePremium}{double}
  \item{RevaluationReserve}{double}
  \item{OtherReserves}{double}
  \item{AccumulatedRetainedEarnings}{double}
  \item{TreasuryStock}{double}
  \item{OtherEquity}{double}
  \item{Equity}{double}
  \item{EquityRatio}{double}
  \item{BalFigCheck}{double}
  \item{NoEmployeesEop}{double}
  \item{NoEmployeesAvg}{double}
  \item{ProductivityPerEmployee}{double}
  \item{AvgEmployeeCost}{double}
  \item{CollectionPeriod}{double}
  \item{InventoryPeriod}{double}
  \item{PayablesPeriod}{double}
  \item{CurrentRatio}{double}
  \item{QuickRatio}{double}
  \item{LiquidAssetsRatio}{double}
  \item{OperatingCycle}{double}
  \item{FinancingGap}{double}
  \item{WorkingCapitalNeed}{double}
  \item{NetSeniorDebtByEbitda}{double}
  \item{InterestCover}{double}
  \item{EBITDAInterestCoverage}{character}
  \item{EstimatedInterestRate}{character}
  \item{AssetTurnover}{double}
  \item{FixedAssetTurnover}{double}
  \item{CurrentAssetTurnover}{double}
  \item{Gearing}{double}
  \item{Leverage}{double}
  \item{STFinancingRatio}{double}
  \item{WkLoansRatio}{double}
  \item{ReturnOnAssets}{double}
  \item{ReturnOnFixedAssets}{double}
  \item{ReturnOnCurrentAssets}{double}
  \item{ReturnOnEquity}{double}
  \item{CFSustTotalSales}{double}
  \item{CFSustEBITDA}{double}
  \item{CFMaintCapex}{double}
  \item{CFInterestExpense}{double}
  \item{CFIncomeTax}{double}
  \item{CFMandatoryDividends}{double}
  \item{CFFreeCashFlow}{double}
  \item{CFMultiplier}{double}
  \item{CFCashFlowValuation}{double}
  \item{CFMaxLeverageST}{double}
  \item{CFMaxLeverageLT}{double}
  \item{CFSeniorDebt}{double}
  \item{CFWkExcessDeficit}{double}
  \item{CFHeadroomAvailable}{double}
  \item{CFHeadroomST}{double}
  \item{CFHeadroomLT}{double}
  \item{CFSeniorDebtByFCF}{double}
  \item{CFHeadroomByFCF}{double}
  \item{EstimatedFreeAssets}{double}
  \item{TotalOverdueDebts}{double}
  \item{TotalOverduePayables}{double}
  \item{OverduePayables1M}{double}
  \item{OverduePayables3M}{double}
  \item{OverduePayables12M}{double}
  \item{OverdueTaxDebts}{double}
  \item{OverdueDebtsOther}{double}
  \item{iTotalOverdueReceivables}{double}
  \item{TotalOverdueReceivables}{double}
  \item{OverdueReceivablesBacklog}{double}
  \item{iOverdueReceivablesBudget}{double}
  \item{OverdueTaxReceivables}{double}
  \item{LTReceivables}{double}
  \item{UnsaleableInventory}{double}
  \item{ShareOverdueDebts}{double}
  \item{ShareOverduePayables}{double}
  \item{ShareOverdueReceivables}{double}
  \item{ShareUnsaleableInventory}{double}
}}
\usage{
Financials2018
}
\description{
Financials 2018 data with 165 variables
}
\keyword{datasets}
