% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{InsolvencyRegistry}
\alias{InsolvencyRegistry}
\title{InsolvencyRegistry data with 5 variables}
\format{A data frame with 32107 rows and 5 variables

\describe{
  \item{ID}{integer}
  \item{FirstRegistrationAge}{double}
  \item{LastRegistrationAge}{double}
  \item{TotalNoRecords}{double}
  \item{AverageNoRecords}{double}
}}
\usage{
InsolvencyRegistry
}
\description{
InsolvencyRegistry data with 5 variables
}
\keyword{datasets}
