% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MovablesRegistry}
\alias{MovablesRegistry}
\title{MovablesRegistry data with 13 variables}
\format{A data frame with 5863 rows and 13 variables

\describe{
  \item{ID}{double}
  \item{NoCreditors}{double}
  \item{OldestRecord}{double}
  \item{NewestRecord}{double}
  \item{NoRecords}{double}
  \item{NoActiveCreditors}{double}
  \item{OldestActiveRecord}{double}
  \item{NewestActiveRecord}{double}
  \item{NoActiveRecords}{double}
  \item{AvgRecords}{double}
  \item{AvgRecordsAge}{double}
  \item{AvgActiveRecords}{double}
  \item{AvgActiveRecordsAge}{double}
}}
\usage{
MovablesRegistry
}
\description{
MovablesRegistry data with 13 variables
}
\keyword{datasets}
