% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_kyuc.R
\name{normalization_kyuc}
\alias{normalization_kyuc}
\title{normalization_kyuc}
\usage{
normalization_kyuc(x)
}
\value{
Normalized vector \code{x}
}
\description{
Normalization is a popular tool, which ensure same range [0,1] for all numerical variables.For further details see \url{https://towardsdatascience.com/normalization-vs-standardization-quantitative-analysis-a91e8a79cebf}.
}
\examples{
example_vector=c(1,2,3,4,5,6,7,8,9,10)
normalization_kyuc(example_vector)
}
