% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization_kyuc.R
\name{standardization}
\alias{standardization}
\title{standardization}
\usage{
standardization(x)
}
\arguments{
\item{x}{a input vector}
}
\value{
standardized vector \code{x}
}
\description{
Standardization is a popular tool for data transformation. It should ensure zero mean and unit variance. For further details see \url{https://towardsdatascience.com/normalization-vs-standardization-quantitative-analysis-a91e8a79cebf}.
}
\examples{
example_vector=c(1,2,3,4,5,6,7,8,9,10)
standardization(example_vector)
}
