% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxOverdues}
\alias{TaxOverdues}
\title{Tax Overdues data with 8 variables}
\format{A data frame with 20004 rows and 8 variables

\describe{
  \item{Year}{integer}
  \item{MainOustandingTaxLiab}{double}
  \item{SecondaryTaxLiab}{double}
  \item{ContestedTaxLiab}{double}
  \item{TotalTaxLiab}{double}
  \item{UncontestedTaxLiab}{double}
  \item{BudgetType}{character}
  \item{ID}{double}
}}
\usage{
TaxOverdues
}
\description{
Tax Overdues data with 8 variables
}
\keyword{datasets}
