% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VATStatus}
\alias{VATStatus}
\title{VAT Status data with 13 variables}
\format{A data frame with 20000 rows and 13 variables

\describe{
  \item{ID}{integer}
  \item{VATStatusAge}{double}
  \item{VATStartAge}{double}
  \item{VATEndAge}{double}
  \item{VATCollectionStartAge}{double}
  \item{VATCollectionEndAge}{double}
  \item{VATCollectionUpdateAge}{double}
  \item{VATCollectionPublishAge}{double}
  \item{VATInactivationAge}{double}
  \item{VATReactivationAge}{double}
  \item{VATCancellationAge}{logical}
  \item{VATSplitStartAge}{double}
  \item{VATSplitCancellationAge}{logical}
}}
\usage{
VATStatus
}
\description{
VAT Status data with 13 variables
}
\keyword{datasets}
