% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize_kyuc.R
\name{winsorization}
\alias{winsorization}
\title{winsorization}
\usage{
winsorization(x, perc_d, perc_u)
}
\arguments{
\item{x}{a input vector}

\item{perc_d}{lower percentile, which will be used for winsoriation}

\item{perc_u}{lower percentile, which will be used for winsoriation}
}
\value{
Winsorized  vector \code{x}
}
\description{
Its purposes is to eliminate outliers in a following way. Values lower or higher than a certain percentile
will be replaced by the percentile. The function is taken from package DescTools.
}
\examples{
example_vector=c(-1000,1,2,3,4,5,6,7,8,9,1000)
winsorization(example_vector,0.05,0.95)
}
